Split as client repo, guess at instructions.
Originally the server and simple client scripts lived in the same
repo, but it would be nice to make it easy to grab just the client,
so here’s this new repo.

I have attempted to write up instructions on how to use it. I still
need to actually go through them myself and stuff…